+++ src/service/resale/factories/make-register-service.ts
@@ -1,9 +1,12 @@
-import { PrismaResaleRepository } from "../../../respositories/prisma/prisma-resale-repository";
+import { PrismaResaleRepository } from "../../../respositories/prisma/prisma-resales-repository";
+import { PrismaUsersRepository } from "../../../respositories/prisma/prisma-users-repository";
 import { RegisterService } from "../register-service";
 
 export function makeRegisterService() {
   const resalesRepository = new PrismaResaleRepository();
-  const service = new RegisterService(resalesRepository);
+  const usersRepository = new PrismaUsersRepository();
+
+  const service = new RegisterService(resalesRepository, usersRepository);
 
   return service;
 }
